Move SRO form to profile
