+++ src/app/[locale]/(logged-in)/organisation/profile/components/NameAndAddress/SroDeclarationUploader.tsx
@@ -50,8 +50,8 @@ const SroDeclarationUploader = ({
   };
 
   return (
-    <Grid container item spacing={3} sx={{ mt: 1 }}>
-      <Grid item xs={4}>
+    <Grid container spacing={3} sx={{ mt: 1 }}>
+      <Grid item xs={5}>
         <FileLink
           fileButtonText={file?.name ? t("buttonTextAlt") : t("buttonText")}
           message={`${FileType.CERTIFICATION}${name.toUpperCase()}`}
