+++ src/app/[locale]/(logged-in)/organisation/profile/components/SroFields/index.ts
@@ -1,0 +1,3 @@
+import SroFields from "./SroFields";
+
+export default SroFields;
