+++ src/config/locales/en.json
@@ -312,7 +312,7 @@
     "detailsOrganisationSectorDescription": "Your Organisation’s sector information should align with Standard Industry Classification (SIC) codes as listed in Companies House or other regulatory sources.",
     "detailsSubsidiaries": "UK subsidiaries",
     "detailsSecurityCompliance": "Data security compliance",
-    "userAdminDelegates": "SRO and Delegates",
+    "userAdminDelegates": "Delegates",
     "userAdministrationTitle": "Users and Delegates",
     "userAdminEmployeeStudents": "Users",
     "manageDelegates": "Manage Delegates",
